--- a/changelog/new_add_new_performance_concurrent_monotonic_time_cop.md
+++ b/changelog/new_add_new_performance_concurrent_monotonic_time_cop.md
@@ -1,0 +1,1 @@
+* [#267](https://github.com/rubocop/rubocop-performance/pull/267): Add new `Performance/ConcurrentMonotonicTime` cop. ([@koic][])